Recompute hash stats for new core vocab
This updates the sample output comment in the bottom of
hash_stats.c with recomputed hashes to reflect the new hashmap
related words that I just added to kernel.mkb. I tried changing the
hash to use the new top-ranking set of parameters for the 64 bin
hash. But, it actually made the code run slightly slower. I'm
guessing that was because different combinations of collisions
resulted in more cycles getting spent on list searching for
frequently used words. So, I left the hash parameters as they were.