Streamline frequently used code, reduce stack use
In the process of trying to get the kernel to compile itself, I've
been using the debug trace feature, which is pretty noisy. I also
noticed that some of the code was using kind of a lot of stack
slots. This change simplifies whitespace detection to reduce clock
cycles spent on parsing words, and it also reduces the stack space
needed for parsing words.