Fix bug with comments in `(` during self-hosting
This was a weird bug, but it kinda makes sense in retrospect. The
definition for `(` had comments in it, so while the definition was
in progress, the self-hosting compiler treated instances of `(` as
recursive calls to `(`. That's wrong because `(` is supposed to be
immediate. But, of course, as the definition is still in progress,
it cannot behave as an immediate word yet.

Before that definition, comments would be handled by the `(` word
that was compiled into the kernel. After the definition, the `(`
new word was marked as immediate and gets used for future comments.
But, at that point, the code for `(` was a bunch of garbage because
it got corrupted by mis-compiling its own comments.

I think this is the issue that motivated the `SMUDGE` word in
figForth. Apparently, another way to resolve the problem is
postponing the adding of words to the dictionary until the
definition is complete. But, that would prevent tail recursion.