Add Pd connection (with patch) to the irc bridge
So, this pretty much just works now. I can type stuff in irssi and
watch it show up as objects and connections in the Pd GUI. I'm
kinda shocked at how straightforward this turns out to be. The
trick is to connect output from the `netreceive` object to a `send`
object with a target of either `pd` or `pd-foo`. The plain `pd`
target allows for stuff like `dsp 1;` or `dsp 0;` to turn the dac
on and off. Targets with a `pd-` prefix send messages to subcanvas
objects created with, for example, `pd foo` in an object box. The
subcanvases will take messages to create and connect objects. You
can type in irc, and they just appear in Pd. It's like magic!