Add auto-retry to pdbridge's connection to Pd
This modifies the Pd side of the pd bridge so it will detect when
the connection goes down and attempt to automatically re-connect
after a short delay. There are lots of reasons why the bot might
not be able to connect to Pd, mostly related to things not being in
the right state with the patch on the Pd side.

I didn't bother with any retry stuff on the irc connection side
since it's meant for use with ngircd on localhost, which should be
rock solid.