Update demo Pd patch for use with pdbridge + irssi
This updates the patch with an additional top level route option
for messages coming out of netreceive. The new behavior is:
1. `pd` prefix -> send to the main pd canvas (`pd`)
2. `send` prefix -> send to the `subsend` receiver in subcanvas
3. the rest -> send to the subcanvas (`pd-sub`)

The main benefits are:
1. I don't need a prefix for messages to create and connect objects
   in the subcanvas
2. I can now send messages to a receive object in the subcanvas,
   rather than just to the subcanvas itself. This lets me do stuff
   like send messages to change the volume

I also made a text file of messages that can be used to re-populate
the subcanvas with a tone generator and volume control. This is
mean to be fed into irssi like so:
   /EXEC -out cat ./join.sh sub-canvas.txt
The join.sh script translates newlines to spaces so the whole file
can be sent as 1 irc PRIVMSG without triggering rate limiting.