Give Markab VM irc powers!
This changes a lot of stuff all at once, but it kinda had to be
this way as the various things are interdependent. There is some
subtlety and weirdness to manage domain-crossing between asyncio
code for irc network stuff and synchronous code for the VM. It's
vaguely like clock-domain crossing in a digital circuit.

The tricky thing is that the VM needs to work with input from a
terminal, from pipes, and from async network code. I'm also trying
to structure the code to keep my options open for maybe running on
MicroPython or CircuitPython down the road. Using async/await code
in the VM class itself would limit my options too much.

Main changes:
1. The kernel rom no longer runs its own input loop. Instead, there
   is a new IRQRX interrupt vector for jumping to a handler that
   interprets one line of text. The code that instantiates a VM is
   responsible for managing an input loop.
2. The VM's standard input and standard output streams are now
   fully line buffered. (used to be byte-buffered)
3. It works over irc! (invoke as `./markab_vm.py --irc`)