Fix VM.print() and format of memory `dump` word
When I tried using irssi to have the Markab VM do a memory dump,
the formatting was messed up. Newlines were missing due to an
oversight in my implementation of VM.print() (irc option assumed
that `end=''` was always provided). Also, the line width was a
little too much to fit on one line after a timestamp and nick.

This commit corrects both problems.