Add string compiling to kernel with `"`
This works to compile strings to the dictionary in both interpret
and compile modes. In compile mode, the string goes inline with its
enclosing word, with a jump at the front and a U16 literal at the
back to push its address. In interpret mode, the string just goes
into the dictionary without any header and its address is left on
the stack.

These new strings use the same string struct format as is already
used in the kernel for `word`, `print`, and `strcmp`. So, they can
be conveniently printed and compared without any further code.

Examples of usage:
```
"   Hey, I'm a string!" print  Hey, I'm a string!  OK
: str "   STR" print ; str  STR  OK
" abc"  " ABC"  strcmp .  0  OK
```