Add `load` (works) and `save` (in progress) words
This extends the VM instructions and kernel words to provide a way
to load and save files. There are two layers of safety precaution
for file paths:
1. The new `IOLOAD` and `IOSAVE` opcodes each have a hardcoded
   allow-list of file path regular expressions, one of which must
   match for a file operation to proceed. For now, load matches
   `.+\.mkb` (like shell glob *.mkb), and save only matches
   `self_hosted\.rom`.
2. Also, the normalized absolute file path must be located in the
   VM's current working directory (including subdirectories).

This is mostly designed to prevent file corruption, or random dumps
to irc, in the event of serious bugs that manage to trigger a file
operation. Using this on a public irc channel is not recommended.

Also, compiling the kernel with `load" kernel.mkb"` seems to work.
The code is not location independent yet, but it seems to run fine
in my limited testing so far. I had to fix a thing with `;`, but
now it's good. Next steps are: finish implementing IOSAVE, make
dictionary links relocatable, and figure out how to self-host the
bootloader.