Rework prompts and kernel error detection
Previously, the kernel was pretty laid back about reporting error
messages. To see some stuff I had to turn on the DEBUG mode. This
commit makes it so that the regular prompt will display error codes
more consistently. But, the big improvement is to display the word,
file, line number, and line text for errors while loading files. It
even works for nested loads. If there are no errors, only one OK
prompt is displayed, once the whole load is done.

The big motivating factor for this was to quiet down the output so
that loading over irc does not produce many hundreds of OK prompts.
Previously I had been relying on line-by-line OK or ? ... prompts
in my terminal scrollback buffer to identify errors. This way is
much better.