Improve prompts for interactive compiling
This modifies the prompting to make it more obvious when:

1. A definition is still in progress (compile mode is active after
   a newline) because the compiler is waiting for some combination
   of `}for`, `}if`, and `;`

2. A definition had an unbalanced arrangement of `}for` or `}if`.
   For example, `: foo }if ;` would earn you an `  ERR 12`

Previously, it was possible to provide input that left compile mode
active without any indication of what was going on. So, it would
look like the interpreter stopped responding to commands. Now there
is a `  Compiling` prompt that takes the place of `  OK` when the
compiler wants you to finish a definition.