Improve error unwinding behavior and prompts
This makes it so the VM will print the "  ERR <n>" prompt on its
own if it stops the code to break what it thinks is an infinite
loop. The `outer` word also does smarter error recovery if it sees
that an error condition is already set when it is invoked.

Overall result is to make sure that an invocation of the outer
interpreter will consistently end with one error or okay prompt and
a newline sequence. Previously, some errors could prevent any
prompt or line ending, and other errors could result in a double
prompt (error followed by okay). This will hopefully prevent weird
missing or extra prompts.