Write Markab code for a Pd tone generator
This code generates Pd messages to build the same tone generator
that I made last week with pdbot. So far, I can ask mkbot to load
this and they give me back text that looks about right. But, I
can't test this for real yet since I still need to build a way for
mkbot to talk directly to pdbot.