Add checks to prevent array out of bounds in VM
There were several opcodes that did not fully check the values they
were using to index into the array that backs the VM's memory. For
example, running `0 here - call` could crash the VM with an array
out of bounds exception. Mostly, the problem was with opcodes that
used PC-relative addressing or read an operand value after the PC.
Hopefully this commit fixes that whole category of bugs.