Start a standard library for utility code
This starts stdlib.mkb with words to check free memory and reboot
the rom. For the moment, this is a way to work around the fact that
I haven't bothered to make the bootstrap compiler smart enough to
properly interpret words at compile time.

The bootstrap compiler is currently hardcoded with implementations
for a core set of immediate words. Those core words are enough to
compile the kernel, but less than the full set of kernel features.
When the bootstrap compiler compiles new immediate words, they just
go into the rom image without becoming available for use during the
bootstrap compile. Maybe I should fix that, but his works for now.