Optimize frequently called VM functions
Compiling the kernel was taking about 6.5 seconds. For this commit,
I used cProfile to identify and optimize some of the VM functions
with the highest cumulative runtime. These changes take the kernel
compile time down to about 6.0 seconds.

Also, the compile times are higher than they would be on a typical
fancy dev laptop. I'm using a low-powered Intel Atom CPU so that
inefficient code will be obvious and annoying, prompting me to work
on speeding it up.

Stuff that seems to help the most:
1. Eliminate extra function calls (makes a big difference)
2. Minimize loads or stores to instance variables like self.DSDeep