Refactor VM: move irc stuff and re-work imports
I kinda want to try running the VM on CircuitPython with IO over
USB serial. With that in mind...

This commit streamlines imports for the VM in the case where it's
being started in terminal IO mode without any profiling or irc
features turned on. I removed the readline dependency and made the
cProfile import conditional.

For the irc stuff, previously it was split between markab_vm.py and
mkb_irc.py. Now all the remaining irc stuff and the asyncio import
have been moved to mkb_irc.py.