Big VM go-faster refactor: change class to globals
This reduces the kernel compile time by approximately 800ms, from
about 5.8 seconds down to about 5.0 seconds.

Previously, the VM was a class, and virtual CPU and RAM state were
stored as class instance variables. Now, the VM is a module, with
all the state stored as global variables of the module.

This change means it's only possible to use one instance of the VM
in a given program. But, one is enough, so that's fine. The benefit
is that I get to save all the overhead for passing arround `self`
pointers as function arguments and for accessing instance variables
through `self`. People talk trash about globals, but the reality
is, they're fast.