Add interrupt vector to handle errors
This change makes it so that an error during an instruction will
trap and cause the CPU to follow the IRQERR interrupt vector to an
error handling routine.

Previously, I allowed instructions to continue in their normal flow
after most errors so the kernel could poll the error register. But,
today I realized that could result in some fairly unpleasant
undefined behavior, particularly for errors in jumps or calls.

This commit also includes an attempt to streamline the non-debug
version of the markab_vm.VM._step() inner loop by refactoring the
conditional debug tracing stuff into a different function. I'm not
sure how much of a net difference there is because I also added
more code to that loop for dealing with the error interrupts.