Make error traps more consistent and efficient
Previously, error detection and recovery was ad hoc and a bit
sloppy, with redundant tests and resets in some cases. The timing
for jumps to the error handler was inconsistent. Sometimes it would
happen on the next clock cycle, but in other cases some additional
instructions ran before the error would get noticed.

This commit streamlines all that to get rid of redundant checks and
to ensure that the program counter will jump to the kernel's error
handler on the next clock cycle after an error. Some of my earliest
tests were built for lazy error polling, so I had to update that
stuff. In particular, I moved the return stack tests from the VM
test set to the kernel rom test set.

Also, the code in this commit clocks in at 4.5 seconds for a kernel
build. So, that's about 500ms better than the last time I checked.