Refactor pdbot as first step to 3-way conversation
I'm trying to figure out how it should work for me to have a 3-way
conversation with mkbot and pdbot. Having separate pd and mkb
channels feels weird. I think it would be better to have everybody
on mkb listening for messages addressing them by name, then we can
use queries for one on one conversations.

This commit includes some filtering so pdbot will respond to either
general greetings or being directly addressed in the main channel.
But, in a query, pdbot will assume all messages other than
greetings are meant for them. This provides a path to experiment
with messages to Pd without having to always write the direct
address prefix at the start of every line.