Get tonegen.mkb working with new channel setup
This commit modifies my old tone generator script to prefix each of
the lines it sends with a mention of pd (pdbot's nick). So, it is
now possible to send these messages in the mkb channel to ask mkbot
to ask pdbot to re-create the tone generator:
```
mk, load" mkbot/tonegen.mkb"
mk, go
```
In response, mkbot will send messages to the mkb channel, addressed
to pdbot, to clear the old tone generator and make a new one.