Generate patch for wavetable oscillator
This adds markab code so mkbot can generate Pd messages to create a
wavetable oscillator with an editable waveform graph.

It took me a long time to figure out how to send a sequence of Pd
messages to approximate Pd's Put:Array menu option to make a graph
widget with an editable array. The trick is to first make a `pd`
box (subpatch), then send it a `coords` message to configure the
graph-on-parent option, then add an array to the subpatch by
sending it an `array ...` message.