First attempt at a kick drum patch
This uses three array graphs to do envelopes for a kick drum patch.
It's kinda passable, maybe? Not terrible. Not fantastic. I would be
happier if I built something to reset the random seed for the noise
generator at the start of each kick.

Also, I added a VU meter to the DAC module and refactored some of
the array graph code over to libpd.mkb.