Refactor subpatch message routing
This frees up space in my main subpatch by moving some of the
message routing to the top level net-listener patch. Also, I shrunk
the volume and dsp control buttons into a narrow-ish column instead
of a fairly large block.

The point is to minimize clutter in the main subpatch. I'm trying
to work within a maximum width of 650px so screenshots will fit in
posts on my dev log without scaling.