Rework file IO to use FOPEN, FWRITE, etc.
This removes the IOSAVE/save/save" stuff from my last commit and
replaces it with wrappers for the host os fopen, fread, fwrite,
fseek, ftell, ftrunc, and fclose file IO functions.

The point of this is to allow for random read/write access to files
which may be bigger than the Markab VM's RAM size. I'm expecting to
use this to build animated GIFs, and perhaps audio samples.