Make dictionaries relocatable
This converts the linked lists for the dictionary hashmap bins to
use relative addresses. After these changes, the link field of an
item in the dictionary contains a modulo-65536 offset, from the
link field's absolute address, to the link target's absolute
address. The code for inserting new entries into the dictionary
(`create`, as called by `:`) knows how to rebase the relative links
when inserting new entries, so it all works out.

These changes make it so the Markab bootstrap compiler and kernel
both emit the functional equivalent of relocatable object code for
the Markab VM.

The point of all this is to facilitate self-hosted recompilation
of the kernel.