Re-work for-loop counter implementation
This changes the `for{...}for` loop counter branching test to use
1-indexed counting rather than 0-indexed counting. The point is to
make it less confusing to specify how many iterations you want the
loop to run for.

Previously, I was using 0-indexed counting and starting almost all
my loops with the `1- for{` idiom. Now, after changing the loop
branching test from R >= 0 to R > 0, I can specify an initial loop
counter value that is equal to the number of iterations I actually
want, and I can omit the `1-`.

For the moment, for loops will always run at least once since the
branching test is at the end of the loop. I'm considering whether
I should add another branching test at the start of the loop to
skip the initial iteration if the loop counter's initial value is
less than 1.