Working on self-hosted kernel compile
This is work in progress. I can compile the kernel and write it out
to a rom file. But, the dictionary hashmap and code for self-hosted
compiled words do not match the output of the bootstrap compiler.
Also, the resulting self-hosted rom file doesn't work. It will run,
but it doesn't seem to be listening to input. There are no errors,
but there are also none of the expected responses.

I think I need to build a rom file disassembler so I can understand
what's going on. Looking at diffs of hexdumps gives me a general
impression, but disassembling by hand is slow and confusing. I
think maybe there's a bug in `find` or `create`. When I manually
test stuff like `load kernel.mkb ' T_VAR . .`, the address that
comes back is from the core vocab when it should be finding T_VAR
in the ext. vocab. Some other stuff works though, like `' here . .`
is fine. Something is weird with the dictionary links maybe. Also,
there may be a problem with optimization in the kernel's compiler
compared to the bootstrap compiler. The bootstrap compiler makes
smaller code.