Start new game engine experiment
This copies the kernel, tests, and syntax highlighting stuff from
repl2 into a new game engine directory. I plan to make a new VM in
C so that I can build it as a WebAssembly module to use with web
APIs (perhaps also SDL?) for graphics, sound, and input devices.