Generate libmkb file structure and boilerplate
This sets things up to compile libmkb as a single translation unit
by using #include with .c files. The point is to have the compiler
able to do inlining optimizations for the bytecode interpreter's
for loop and switch statement, which will be a very hot code path.

Using autogen for the big switch statement allows the option to
conveniently add or remove opcodes in the future. Otherwise, I
would have to hand update the numbers in all the cases, which
would totally suck. But, autogen won't work for the opcode
implementation functions because I manually edit those. So, the
opcode functions have to go in another file. The solution is to use
`#include` with .c files.