Reduce Plan9 vs POSIX #include chaos
Previously, I had #includes of header files from inside of header
files. That was a bad idea, so I stopped. This new arrangement
leaves it up to the .c files to include the right headers in the
right order (like how Plan 9 C compilers need #include <u.h>).

Also, I've changed the 9front mkfile with a `cd libmkb; ...` so the
libmkb object file gets compiled in engine/libmkb instead of in
engine/. This solves the problem with 6c and clang handling include
paths differently. Now, all my includes refer to header files in
the current directory, which both compilers are happy with.