Improve macros, port load/store and literals
This is a big messy commit. The work here started off with just
implementing the literal-loading and load/store opcodes. To do
that, I needed to write some peek and poke macros. That grew into
a big tangent for cleaning up macro stuff in general.

What I'm doing with the macros is basically making a little DSL for
a lot of code fragments that I need frequently for implementing
opcodes. I'm trying to make the code more readable by using
meaningful macro names in place of repetitive chunks of code
that would otherwise be pretty dense and confusing.

I could use function calls for some of this. Perhaps I should? For
now, I want to continue with macros. I'm liking how the readability
feels so far.