Remove all file IO opcodes (for now)
To simplify the process of getting this new C VM up and running
with the kernel and tests from my previous Python VM, I'm going to
take file IO out of the back end library code for the moment. Now,
the front end code will be responsible for providing the VM with
stdin and stdout streams.

I'm also re-working my architecture to be more of a bytecode
intepreter and less of a virtual CPU. For example, I had been
limiting virtual CPU opcodes to the types of things that might
reasonably be implemented in an FPGA some day. Now, I don't think
that arbitrary self-imposed constraint of trying to emulate a
hardware device is serving me well at all.

Refactoring some compute-intensive tasks from Markab kernel words
to become opcodes implemented in C will make this whole thing lots
more efficient. In particular, `find` uses a ton of virtual CPU
operations each time it's invoked, and that seriously clutters up
debug traces. Moving `find` out of the kernel should help with
debugging Markab code because the traces will be shorter.