Implement DOT opcode, add HEX and DECIMAL
This provides destructive hexadecimal or decimal format printing of
the top data stack item. Previously, I did number formatting and
parsing for DOT (formerly IODOT) in the markab kernel, so the base
was kept in kernel memory space.

By adding HEX and DECIMAL opcodes, I'm getting more serious about
moving number formatting responsibility to opcode implementations.
This is part of my plan to refactor a lot (all?) of the compiler to
be implemented in C rather than markab. Instead of having a
bootstrap compiler and another interpreted compiler as part of the
kernel, I want to have just one compiler, written in C.