Remove compiler-related opcode & mem-map stuff
This strips out a bunch of opcode and memory-map stuff that I was
using for string parsing from within the markab kernel with my old
Python VM.

The motivating problem here is that it's confusing and slow to
implement string copying and comparison algorithms as interpreted
bytecode loops. Before, when I was focused on build a self-hosting
kernel and compiler, tolerating that pain was necessary. But, now I
can just move all the compiler's string parsing stuff to C code.

The general idea is that I want the markab language to become less
like a Chuck-Moore-hardware-Forth-ish thing and become more like a
friendly-to-use-Forth-meets-HyperScript-ish thing. To the extent I
find that I need string processing operations in markab scripts, I
can implement opcodes in C to do the heavy lifting.