More Tests! Tests are good.
This sets up a minimalist but actually-not-too-terrible test runner
that keeps score of pass/fail results and returns a proper exit
status. It works on Debian, macOS, and Plan 9.

In the course of implementing this, I discovered that my decimal
integer formatter was mishandling -2,147,483,648. Seems like Plan 9
is pretty good for catching weird edge-case bugs that clang doesn't
necessarily complain about. Anyhow, I fixed the decimal formatter.