Finish tests for BZ and JMP opcodes
I had some trouble with the address calculation for jumping to a
negative PC-relative offset. So, in the course of sorting that out,
I made some improvements to error message displays and a couple
other things that I thought might possibly have been related to the
problem. Turned out the issue was using u32 in a struct when I
really, really should have been using u16.