Fix snprintf warnings on Debian
So, snprintf() is actually a C99 feature. But, `clang -ansi` on
macOS and 6c on Plan 9 (ANSI C + extensions) both allow snprintf().
That's good, since it's much safer than old school sprintf(). But,
on Debian 11, `clang -ansi` gives a warning.

This change fixes the warning by adding an _XOPEN_SOURCE define
when clang detects that it's not running on macOS.