Add STR and PRINT opcodes
STR is for encoding counted string literals. It works like U8, U16,
and I32, sorta. PRINT is for printing a counted string, including,
but not limited to, string literals encoded in a STR.

My main goal with STR and PRINT is to alleviate the current hassle
of using my `MK_U8, 'c', MK_EMIT,` idom to print strings one
character at a time. The new opcodes will make it much less tedious
to hand-compile strings for test code.