Finish tests for DROP & DUP, fix vm_irq_err()
In the course of testing to make sure DROP would give a stack
underflow error, I realized that vm_irq_err() was broken. It was
only logging errors and deciding on whether to halt the VM. But,
it should also have been remembering the error code so that
mk_load_rom() could return it to the front end code. Fixed that.