Finish tests for arithmetic ops, fix DIV & MOD
This finishes tests for ADD, SUB, MUL, DIV, and MOD. The DIV and
MOD opcodes both had two potentially process-killing bugs. To avoid
the possibility of the host OS killing the bytecode interpreter's
process, DIV and MOD now validate the dividend and divisor.

Dividing anything by zero is an error. But, also, the CPU can make
a hardware trap if the quotient overflows its destination register.
For signed 32-bit integer operations (-2147483648 / -1) or
(-2147483648 % -1), the quotient of 2147483648 is bigger than the
maximum positive value that fits in a int32_t: 2147483647.

So, I fixed DIV and MOD to check for divide by zero and quotient
overflow errors before attempting the division operation.

ADD, SUB, and MUL are subject to overflow conditions, but not in
a way that triggers a CPU trap. Intentional overflows can be
useful, so I don't attempt to prevent them.