Prune unused codegen stuff
This gets rid of some more dead code that I've been carrying along
since I copied it over from my previous VM implemented in Python.

This stuff I removed was meant to support a kernel and compiler
implemented as Markab bytecode. Now that I'm moving all those
features to C-land, the old codegen stuff isn't so suitable. Also,
I want to re-do the codegen in C, so this simplifies the amount of
work to port it.