Rename word equivalents of INC and DEC opcodes
This changes the INC word from 1+ to ++ and the DEC word from 1-
to --

My goal here is to simplify parsing by imposing the rule that any
lexical token starting with 0-9 in the leading byte gets treated as
a number. I still need to fix 0= for the ZE opcode, but that's kind
of a bigger deal, so I'll save it for another commit.