Fix int literals + add compiler error messages
Parsing of int literals seems to be working well now. I added a
compiler error message printing mechanism that includes the text
which triggered the error. There's even a caret pointing to the
column where the problem was detected. Compared to error output
from my previous experiments, this is feeling downright luxurious.