Make triangle strips for entire screen
This generates triangle strips to tile a 240x160 pixel screen with
10 rows and 15 columns of 16x16 squares, two triangles per square.

Rather than getting fancy by detecting the canvas width and height,
then using that to transform from px coordinates to clip space
coordinates, I just hardcoded the transform for 240x160 into the
vertex shader.

This is probably not optimally efficient because the CSS styles add
another 2x zoom to scale the canvas up to 480x320. For now, I'm not
going to worry about that. Maybe later.