Optimize page load for less jank
Previously, all of the WebGL shader compiling and program linking
happened when the main.js module loaded. This commit factors the
WebGL and Wasm initialization into several functions that get
spread across a chain of requestAnimationFrame callbacks.

There's also a new `PAUSED` interlock that keeps the main animation
loop from starting until a gamepadconnected event happens.

Point of all of this is to make the initial page load happen very
quickly without loading down the main UI thread. It was already
reasonably fast, but what I'm aiming for here is to be able to run
okay on firefox-esr on a Pi 3. I want the CPU load to be small.