Adjust CSS, shaders, etc. to reduce GPU load
This makes it so all the scaling and drawing for each frame pretty
much happens in one pass over the framebuffer. Previously, I had a
redundant (kinda) clear and a CSS scaling step that got applied to
the output of the fragment shader. So, I think it was doing three
passes per frame?

Goal of this is to get a sense of what my baseline CPU and GPU load
look like when there's a minimum of extra nonsense going on.