Debounce dpad buttons for more precise diagonals
This sets up a little state machine to keep track of dpad buttons
so that it's easier to start moving on a precise diagonal. Also,
the debounce delay allows for a rapid press and relase to change
direction without taking a step. I don't have a visual way to
indicate that yet, but I think the code is probably okay.

Also, I adjusted the repeat rate on diagonals so that they don't
appear to move faster than regular right-angles motion.

I also cleaned out some dead code related to the old raster frame
buffer. I may bring something similar back to use with generating
textures in markabscript from inside the wasm module. In that case,
I'll just go back through the commit history for a copy'n'paste.