Start a second game engine directory
I'm moving the previous contents of engine/ to old_engine/, then
starting the new version of engine/ off with a copy of *only* the
Wasm and HTML/JS/CSS stuff from the previous experiment.

The point of this is to set aside the previous engine/libmkb/ code
that implemented a Markab Script compiler. I think that whole VM
and bytecode thing is not actually going to help me do what I want
to do next. I may end up using some of the parsing and compiler
code as part of a level editor... maybe. But, if so, I'll do the
copy and paste later.