Change new engine license to CC BY-NC-SA
Since I plan to start encoding textures, geometry, and so on as C
source code, I don't think continuing under the MIT license would
be the right choice.

The rest of the code for previous experiments is still MIT. In case
of any doubts, look for a copyright notice in the SPDX header
comments, image metadata, etc.